+++ controllers/transactionController.js
@@ -1,4 +1,7 @@
 const { db } = require('../database/index');
+const fs = require('fs');
+
+const { uploader } = require('../helper/uploader');
 
 module.exports = {
   getTransaction: (req, res) => {
@@ -31,7 +34,53 @@ module.exports = {
       res.status(200).send(dataResult);
     });
   },
+  // add custom order transaction
+  addCustomTransaction: (req, res) => {
+    try {
+      let path = '/images/transaction';
+      // TRS for custom order image
+      const upload = uploader(path, 'TRS').fields([{ name: 'file' }]);
 
+      upload(req, res, (error) => {
+        if (error) {
+          console.log(error);
+          res.status(500).send(error);
+        }
+        // if isnt error
+        const { file } = req.files;
+        const filepath = file ? path + '/' + file[0].filename : null;
+
+        // parsing the data
+        let data = JSON.parse(req.body.data);
+        data.prescription_img = filepath;
+
+        let addNewTransactionData = `INSERT INTO transactions VALUES (null, ${db.escape(
+          data.userId
+        )},  ${db.escape(data.recipent)}, ${db.escape(
+          data.address
+        )}, null, "custom", ${db.escape(
+          data.transaction_date
+        )}, null, "unpaid", ${db.escape(filepath)}, ${db.escape(
+          data.notes
+        )}, null);`;
+
+        db.query(addNewTransactionData, (error, results) => {
+          if (error) {
+            console.log(error);
+            fs.unlinkSync('./public' + filepath);
+            res.status(500).send(error);
+          }
+          res.status(200).send({
+            message:
+              'Your query has succesfully been sent, please wait and check your notification for our reply!',
+          });
+        });
+      });
+    } catch (error) {
+      console.log(error);
+      res.status(500).send(error);
+    }
+  },
   addTransaction: (req, res) => {
     let {
       userId,
@@ -122,7 +171,7 @@ module.exports = {
             dataUpdate.push(`${prop} = ${db.escape(data[prop])}`);
           }
 
-          let updateTransactionDataQuery = `UPDATE sys.transaction SET ${dataUpdate} WHERE transaction_id = ${req.params.transaction_id};`;
+          let updateTransactionDataQuery = `UPDATE transactions SET ${dataUpdate} WHERE transaction_id = ${req.params.transaction_id};`;
 
           db.query(updateTransactionDataQuery, (err, results) => {
             if (err) {
