+++ controllers/transactionControllers.js
@@ -33,54 +33,54 @@ module.exports = {
       res.status(200).send(results);
     });
   },
-  // add new transaction Data
-  addNewTransactionData: (req, res) => {
-    try {
-      let path = '/images/transaction';
+  // add new CUSTOM ORDER transaction Data
+  // addNewTransactionData: (req, res) => {
+  //   try {
+  //     let path = '/images/transaction';
 
-      const upload = uploader(path, 'TRS').fields([{ name: 'file' }]);
+  //     const upload = uploader(path, 'TRS').fields([{ name: 'file' }]);
 
-      upload(req, res, (error) => {
-        if (error) {
-          console.log(error);
-          res.status(500).send(error);
-        }
-        // if isnt error
-        const { file } = req.files;
-        const filepath = file ? path + '/' + file[0].filename : null;
+  //     upload(req, res, (error) => {
+  //       if (error) {
+  //         console.log(error);
+  //         res.status(500).send(error);
+  //       }
+  //       // if isnt error
+  //       const { file } = req.files;
+  //       const filepath = file ? path + '/' + file[0].filename : null;
 
-        // parsing the data
-        let data = JSON.parse(req.body.data);
-        data.product_img = filepath;
+  //       // parsing the data
+  //       let data = JSON.parse(req.body.data);
+  //       data.product_img = filepath;
 
-        let addNewTransactionData = `INSERT INTO transaction_details VALUES (null,
-        ${db.escape(data.product_name)}, ${db.escape(data.product_desc)}, 
-        ${db.escape(filepath)}, ${db.escape(data.stock)}, 
-        ${db.escape(data.netto)}, 
-        ${db.escape(data.netto_total)}, ${db.escape(data.unit)}, 
-        ${db.escape(data.price_per_unit)}, 
-        ${db.escape(data.price_per_stock)}, 
-        ${db.escape(data.brand_id)}, 
-        ${db.escape(data.category_id)}, 
-        ${db.escape(data.notes)});`;
+  //       let addNewTransactionData = `INSERT INTO transaction_details VALUES (null,
+  //       ${db.escape(data.product_name)}, ${db.escape(data.product_desc)},
+  //       ${db.escape(filepath)}, ${db.escape(data.stock)},
+  //       ${db.escape(data.netto)},
+  //       ${db.escape(data.netto_total)}, ${db.escape(data.unit)},
+  //       ${db.escape(data.price_per_unit)},
+  //       ${db.escape(data.price_per_stock)},
+  //       ${db.escape(data.brand_id)},
+  //       ${db.escape(data.category_id)},
+  //       ${db.escape(data.notes)});`;
 
-        db.query(addNewTransactionData, (error, results) => {
-          if (error) {
-            console.log(error);
-            fs.unlinkSync('./public' + filepath);
-            res.status(500).send(error);
-          }
-          res.status(200).send({
-            message:
-              'Your query has succesfully been sent, please wait and check your notification for our reply!',
-          });
-        });
-      });
-    } catch (error) {
-      console.log(error);
-      res.status(500).send(error);
-    }
-  },
+  //       db.query(addNewTransactionData, (error, results) => {
+  //         if (error) {
+  //           console.log(error);
+  //           fs.unlinkSync('./public' + filepath);
+  //           res.status(500).send(error);
+  //         }
+  //         res.status(200).send({
+  //           message:
+  //             'Your query has succesfully been sent, please wait and check your notification for our reply!',
+  //         });
+  //       });
+  //     });
+  //   } catch (error) {
+  //     console.log(error);
+  //     res.status(500).send(error);
+  //   }
+  // },
   // update transaction data
   updateTransactionData: (req, res) => {
     try {
